+++ lib/src/hashlib_base.dart
@@ -20,6 +20,7 @@ export 'poly1305.dart';
 export 'random.dart';
 export 'registry.dart';
 export 'ripemd160.dart';
+export 'ripemd320.dart';
 export 'scrypt.dart';
 export 'sha1.dart';
 export 'sha224.dart';